TEST: AuthControllerTest

+++ app/Http/Controllers/API/AuthController.php
@@ -6,44 +6,56 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 
 class AuthController extends Controller
 {
     public function register(Request $request)
     {
-        $validator = $request->validate([
+        $request->validate([
             'name' => 'required|string|max:255|min:3',
             'email' => 'required|string|email|max:255|unique:users,email',
             'password' => 'required|string|min:6|confirmed',
-            'password_confirmation' => 'required|string|min:6',
         ]);
-        User::create($validator);
-        return response()->json(['message' => 'User created successfully.'],201);
 
+        $user = User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => Hash::make($request->password),
+        ]);
+
+        return response()->json(['message' => 'User created successfully.'], 201);
     }
 
     public function login(Request $request)
     {
-        $validator = $request->validate([
+        $request->validate([
             'email' => 'required|string|email|max:255|exists:users,email',
             'password' => 'required|string|min:6',
         ]);
 
-
         $user = User::where('email', $request->email)->first();
 
         if (!$user || !Hash::check($request->password, $user->password)) {
-            return response()->json(['message' => 'The provided credentials are incorrect.'], 401);
+            throw ValidationException::withMessages([
+                'email' => ['The provided credentials are incorrect.'],
+            ]);
         }
+
         $token = $user->createToken('my-app-token')->plainTextToken;
 
-        return success_response(true);
-
+        return response()->json([
+            'message' => 'Logged in successfully.',
+            'token' => $token,
+        ], 200);
     }
+
     public function logout(Request $request)
     {
-        $request->user()->tokens()->delete();
-        return response()->json(['message' => 'Logged out.'], 200);
+        if ($request->user()) {
+            $request->user()->tokens()->delete();
+            return response()->json(['message' => 'Logged out.'], 200);
+        }
+        return response()->json(['message' => 'User not authenticated.'], 401);
     }
-
 }
